requirements.txt
BeautifulSoup==3.2.1
DateTime==4.0.1
Django==1.7.1
MySQL-python==1.2.5
Unipath==1.0
argparse==1.2.1
django-annoying==0.8.0
django-m2m-history==0.1.2
django-oauth-tokens==0.3.12
django-picklefield==0.3.1
django-taggit==0.12.2
django-twitter-api==0.1.7
httplib2==0.9
oauth2==1.5.211
oauthlib==0.7.2
python-dateutil==2.3
python-openid==2.2.5
python-twitter==2.1
pytz==2014.10
requests==2.5.0
requests-oauthlib==0.4.2
simplejson==3.6.5
six==1.8.0
tweepy==3.1.0
tyoi.OAuth2==0.2.1
wsgiref==0.1.2
zope.interface==4.1.1
